Add API call to run arbitrary volatility plugins
Add the ability to execute an arbitrary volatility plugin on the full
memory dump for the task specified - only a small subset of plugins are
currently mapped. This is useful in cases where the plugin a user may
want to run will depend on the output of the analysis or they want to
get memory dumps for use in IDA or a “high throughput” system where
it’s not always desired to have all plugins run in the reporting stage.

Calls are of the form

    POST /tasks/volatility/<task_id>/<plugin>

    config_name=config_value

e.g. a call to procdump might look like

    POST /tasks/volatility/1/procdump

    memory=True&unsafe=False&pid=1,2,3,4

* Any plugin-specific or volatility specific configuration parameters
should be set in the post body.

* If a plugin such as procdump or dlldump is called, a temporary
volatility directory will be created in the storage path and the
specified plugin will be run with the post parameters as options. Any
dumped executables or dlls will be returned in a zip-file along that
also contains the output log as “volatility.log”

* Any plugin that returns text will currently return the plain-text
output in the response body

@TODO figure out a “clean” way to return JSON output instead of text
output and allow that to be specified